# docs/products/swagger.yaml
definitions:
  response.Meta:
    properties:
      code:
        type: integer
      error:
        type: object
      message:
        type: string
    type: object
  response.Single:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    This is a products server.
    you can access the code on github https://github.com/huzairuje/go-echo-example/
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Products API
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - '*/*'
      description: get List Product
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Single'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Single'
      summary: List Product
      tags:
      - products
    post:
      consumes:
      - '*/*'
      description: Post Create Product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Single'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Single'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Single'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Single'
      summary: Create Product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: get Delete Product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Single'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Single'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Single'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Single'
      summary: Delete Product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: get Detail Product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Single'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Single'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Single'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Single'
      summary: Detail Product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: get Update Product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Single'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Single'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Single'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Single'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Single'
      summary: Update Product
      tags:
      - products
schemes:
- http
swagger: "2.0"